requirements.txt
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
matplotlib==3.8.2
scikit-learn==1.4.0
tqdm==4.66.1
numpy==1.26.4
opencv-python==4.9.0.80
gradio==4.19.2
